Atualizações no Back.js
