+++ project/src/Back.js
@@ -1,10 +1,11 @@
 const http = require('http'); 
 const express = require('express');
 const mysql = require('mysql');
-const validUrl = require('valid-url');
-const fs = require('fs');
 
 const getShortenUrlRoute = require('./rotas/GetShortenUrl');
+const linksRoute = require('./rotas/Links');
+const hitsRoute = require('./rotas/Hits');
+const shortUrlRoute = require('./rotas/ShortUrl');
 
 const app = express();
 app.use(require("cors")());
@@ -38,89 +39,11 @@ connection.connect(function(err) {
     });
 
     connection.end();
-})
- 
-app.get('/links', (req, res, next) => {
-    const select = "SELECT * FROM `links` ORDER BY (`hits`) DESC LIMIT 5;";
-    const connection = startConnection();
-    connection.query(select, function(err, result, fields) {
-        try {
-            if (err) {
-                res.json({ Erro: 'Ocorreu um erro ao buscar dados no Banco de Dados.' });
-                return;
-            }
-    
-            if (result == ''){
-                res.json({ Erro: 'Não há links gerados!' });
-                return;
-            }
-    
-            res.json(result);
-        } finally {
-            connection.end();
-        }
-    });
-});
-
-app.get('/hits', (req, res, next) => {
-    const select = "SELECT SUM(`hits`) AS `hits` FROM `links`";
-    const connection = startConnection();
-    connection.query(select, function(err, result, fields) {
-        try {
-            if (err) return res.json(err);
-            return res.json(result[0].hits);
-        } finally {
-            connection.end();
-        }
-    });
-});
-
-app.get('/shortUrl', async (req, res) => {
-    var url = req.query.url;
-    if (!validUrl.isUri(url))
-        return res.json({ Erro: 'Informe uma URL válida!' });
-        
-    var code = makeCharacters(8);
-    var newUrl = url_shorter + "v1/" + code;
-    var select = "SELECT * FROM `links` WHERE `url`=?;";
-    var insert = "INSERT INTO `links` VALUES(NULL, ?, ?, ?, ?);";
-
-    const connection = startConnection();
-    connection.query(select, [ url ], function(err, result, fields) {
-        try {
-            if (err) return res.json(err);
-            if (result != '') return res.json({ newUrl: result[0].shortUrl });
-            
-            connection.query(insert, [ 0, url, newUrl, code ], function(err, result, fields) {
-                if (err) return res.json(err);
-                res.json({ newUrl: newUrl });
-            });
-        } finally {
-            connection.end();
-        }
-    });
 });
 
 /* Função para iniciar a conexão com o Banco de Dados. */
 function startConnection() {
     return mysql.createConnection({ host: host, port: port, user: user, password: password, database: database  });
-}
-
-/* Função para criar caracteres aleatórios para o encurtador de links */
-function makeCharacters(length) {
-    var result           = '';
-    var characters       = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
-    var charactersLength = characters.length;
-    for ( var i = 0; i < length; i++ ) {
-        result += characters.charAt(Math.floor(Math.random() * charactersLength));
-    }
-
-    return result;
-}
-
-function read(file) {
-    let content = fs.readFileSync(file, "utf-8");
-    return JSON.parse(content);
 }
  
 var server = http.createServer(app); 
@@ -128,3 +51,4 @@ server.listen(3030);
 console.log("Servidor escutando na porta 3030...")
 
 app.use("/v1/", getShortenUrlRoute);
+app.use("/", linksRoute, hitsRoute, shortUrlRoute);
